修改删除申请单同时删除流程在一个事物中

+++ src/main/java/com/hr319wg/xys/workflow/service/ActivitiToolsService.java
@@ -179,7 +179,10 @@ public class ActivitiToolsService {
 	 * @throws SysException
 	 */
 	public void deleteProcessInstance(String instanceId)throws SysException{
-		this.runtimeService.deleteProcessInstance(instanceId, "审批退回删除流程实例");
+		ProcessInstance processInstance =(ProcessInstance)runtimeService.createProcessInstanceQuery().processInstanceId(instanceId).singleResult();
+		if(processInstance!=null){
+			this.runtimeService.deleteProcessInstance(instanceId, "审批退回删除流程实例");			
+		}
 	}
 	
 	/**
